+++ CHANGELOG.md
@@ -4,11 +4,14 @@
 * [[#31](https://github.com/VadimDez/ng2-filter-pipe/issues/31)] - How to filter by two variables of the same array.
 * [[#4](https://github.com/VadimDez/ng2-filter-pipe/issues/4)] - Add $or operator.
 
+## 0.1.9
+* [[#32](https://github.com/VadimDez/ng2-filter-pipe/issues/32)] - Filter fails when filter value is zero
+
 ## 0.1.8
 * Removed warning in Angular 4+
 
 ## 0.1.7
-* [[#28](https://github.com/VadimDez/ng2-filter-pipe/pull/28)] - Filter number by string
+* [[#28](https://github.com/VadimDez/ng2-filter-pipe/issues/28)] - Filter number by string
 * Removed post install script
 
 ## 0.1.6